renamed hashed and added new header for SplitSDKImpressionsMode
