+++ src/impressions/observer/node.js
@@ -1,5 +1,5 @@
 import ImpressionObserver from './observer';
-import { hashImpression128 } from '../hasher/node';
+import { hashImpression128 } from '../hasher/hashImpression128';
 
 const LAST_SEEN_CACHE_SIZE = 500000; // cache up to 500k impression hashes
